worker_agents/ueba_agent/requirements.txt
fastapi
uvicorn
pydantic
python-dateutil
scikit-learn
numpy
